[Tests] JS units: move config tests to each dedicated directory
