+++ tests/js-units/node/config/base.test.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 
-import { ValidationError, ConversionError } from '../../../assets/src/modules/Errors.js';
-import { Extent } from '../../../assets/src/modules/config/Tools.js';
-import { BaseObjectConfig, BaseObjectLayerConfig, BaseObjectLayersConfig } from '../../../assets/src/modules/config/BaseObject.js';
+import { ValidationError, ConversionError } from '../../../../assets/src/modules/Errors.js';
+import { Extent } from '../../../../assets/src/modules/config/Tools.js';
+import { BaseObjectConfig, BaseObjectLayerConfig, BaseObjectLayersConfig } from '../../../../assets/src/modules/config/BaseObject.js';
 
 describe('BaseObjectConfig', function () {
     it('Valid', function () {
