+++ tests/js-units/node/config/layer.test.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 
-import { ValidationError, ConversionError } from '../../../assets/src/modules/Errors.js';
-import { Extent } from '../../../assets/src/modules/config/Tools.js';
-import { LayerConfig, LayersConfig } from '../../../assets/src/modules/config/Layer.js';
+import { ValidationError, ConversionError } from '../../../../assets/src/modules/Errors.js';
+import { Extent } from '../../../../assets/src/modules/config/Tools.js';
+import { LayerConfig, LayersConfig } from '../../../../assets/src/modules/config/Layer.js';
 
 describe('LayerConfig', function () {
     it('Valid', function () {
