+++ tests/js-units/node/config/options.test.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 
-import { ValidationError, ConversionError } from '../../../assets/src/modules/Errors.js';
-import { Extent } from '../../../assets/src/modules/config/Tools.js';
-import { OptionsConfig } from '../../../assets/src/modules/config/Options.js';
+import { ValidationError, ConversionError } from '../../../../assets/src/modules/Errors.js';
+import { Extent } from '../../../../assets/src/modules/config/Tools.js';
+import { OptionsConfig } from '../../../../assets/src/modules/config/Options.js';
 
 describe('OptionsConfig', function () {
     it('Valid', function () {
